Bug fixes
Model binding in the new dialogs was not working because defaults were
no longer set. Defaults are now generated as part of the admin view.

Library was PUTting files to a hard coded base path.

New views are now created when the new button is clicked for both lists
and libraries